Refactor XPath evaluator helpers and context logic
